backend/python/requirements.txt
torch>=2.0.0
timm>=0.9.0
huggingface-hub>=0.19.0
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
